chore: rework policy collection item resource [ENG-5144] (#30)
* chore: rework policy collection item resource [ENG-5144]

[ENG-5144](https://stacklet.atlassian.net/browse/ENG-5144)

### what

move API logic to separate package, reuse common architecture

### why

uniform logic for resources

### testing

unit tests and manual testing

### docs

updated here

* rename to policy_collection_mapping

* switch to upsert

[ENG-XXXX](https://stacklet.atlassian.net/browse/ENG-XXXX)

### what

<here>

### why

<here>

### testing

<here>

### docs

<here>

+++ internal/models/policy.go
@@ -11,7 +11,7 @@ type PolicyDataSource struct {
 	Name            types.String `tfsdk:"name"`
 	Description     types.String `tfsdk:"description"`
 	CloudProvider   types.String `tfsdk:"cloud_provider"`
-	Version         types.Int64  `tfsdk:"version"`
+	Version         types.Int32  `tfsdk:"version"`
 	Category        types.List   `tfsdk:"category"`
 	Mode            types.String `tfsdk:"mode"`
 	ResourceType    types.String `tfsdk:"resource_type"`
